+++ services/api-server/src/simcore_service_api_server/api/root.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter
 
 from ..core.settings import AppSettings
-from .routes import files, health, meta, solvers, users
+from .routes import files, health, jobs, meta, solvers, users
 
 
 def create_router(settings: AppSettings):
@@ -14,4 +14,6 @@ def create_router(settings: AppSettings):
     if settings.beta_features_enabled:
         router.include_router(files.router, tags=["files"], prefix="/files")
         router.include_router(solvers.router, tags=["solvers"], prefix="/solvers")
+        router.include_router(jobs.router, tags=["solvers"], prefix="/solvers")
+
     return router
